Better setup for modern/legacy build - much simpler babel config - separate entry points for modern and legacy (ie11/phantom) - explicitly import 'core-js' and 'regenerator-runtime' instead of using deprecated 'babel/polyfill' plugin - still need 'core-js-bundle' for the test/index.html, so Mocha/PhantomJS combo will work.
