+++ babel.config.json
@@ -1,22 +1,9 @@
 {
-  "presets": [
-    [
-      "@babel/preset-env", {
-        "useBuiltIns": "entry",
-        "corejs": 3,
-        "modules": false,
-        "targets": {
-          "chrome": "58",
-          "ie": 11
-        }
-      }
-    ]
-  ],
+  "compact": false,
   "plugins": [
-    "inline-json-import",
-    ["@babel/transform-runtime", {
-      "regenerator": true
-    }]
+    ["inline-json-import", {}]
   ],
-  "compact": false
+  "presets": [
+    ["@babel/preset-env", {"modules": false, "targets": { "ie": "11" }}]
+  ]
 }
