+++ config/esbuild.config.modern-min.mjs
@@ -5,7 +5,7 @@ esbuild
     minify: true,
     bundle: true,
     sourcemap: true,
-    entryPoints: ['./modules/id.js'],
+    entryPoints: ['./modules/id.modern.js'],
     legalComments: 'none',
     logLevel: 'info',
     outfile: 'dist/iD.min.js'
