+++ package.json
@@ -52,7 +52,7 @@
     "abortcontroller-polyfill": "^1.4.0",
     "aes-js": "^3.1.2",
     "alif-toolkit": "^1.2.9",
-    "core-js-bundle": "^3.19.0",
+    "core-js": "^3.19.0",
     "diacritics": "1.3.0",
     "fast-deep-equal": "~3.1.1",
     "fast-json-stable-stringify": "2.1.0",
@@ -64,12 +64,12 @@
     "pbf": "^3.2.1",
     "polygon-clipping": "~0.15.1",
     "rbush": "3.0.1",
+    "regenerator-runtime": "^0.13.9",
     "whatwg-fetch": "^3.4.1",
     "which-polygon": "2.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.15.8",
-    "@babel/plugin-transform-runtime": "^7.15.8",
     "@babel/preset-env": "^7.15.6",
     "@fortawesome/fontawesome-svg-core": "^1.2.32",
     "@fortawesome/free-brands-svg-icons": "~5.15.1",
@@ -81,6 +81,7 @@
     "babel-plugin-inline-json-import": "^0.3.2",
     "btoa": "^1.2.1",
     "chai": "^4.1.0",
+    "core-js-bundle": "^3.19.0",
     "cldr-core": "37.0.0",
     "cldr-localenames-full": "37.0.0",
     "colors": "^1.1.2",
@@ -96,8 +97,8 @@
     "happen": "^0.3.1",
     "js-yaml": "^4.0.0",
     "json-stringify-pretty-compact": "^3.0.0",
-    "mapillary-js": "4.0.0",
     "mapillary_sprite_source": "^1.8.0",
+    "mapillary-js": "4.0.0",
     "minimist": "^1.2.3",
     "mocha": "^7.0.1",
     "mocha-phantomjs-core": "^2.1.0",
